influencer-pipeline/requirements.txt
numpy
pandas
tqdm
regex
emoji
langdetect
orjson
openai
groq
python-dotenv
google-cloud-storage

